remove old & dup semconv
