+++ instrumentation/http/Appraisals
@@ -4,17 +4,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-# To facilitate HTTP semantic convention stability migration, we are using
-# appraisal to test the different semantic convention modes along with different
-# HTTP gem versions. For more information on the semantic convention modes, see:
-# https://opentelemetry.io/docs/specs/semconv/non-normative/http-migration/
-
-semconv_stability = %w[dup stable old]
-
-semconv_stability.each do |mode|
-  %w[5.0.0 6.0.4].each do |version|
-    appraise "http-#{mode}-#{version}" do
-      gem 'http', "~> #{version}"
-    end
+%w[5.0.0 6.0.4].each do |version|
+  appraise "http-#{mode}-#{version}" do
+    gem 'http', "~> #{version}"
   end
 end
